Mark return value of textilize_without_paragraph HTML safe
In Rails 3.2.2 this line was added:
https://github.com/rails/rails/blob/v3.2.2/activesupport/lib/active_support/core_ext/string/output_safety.rb#L102

ActiveSupport::SafeBuffer#[] with 0 or 1 argument turns the buffer into a normal non-safe string. In the helper we pass one argument (a range) to slice away the <p> tags, thus ending up returning an unsafe string as opposed to earlier where this guard didn't exist.